Fix serious bug.

+++ src/main/java/conglin/clrpc/transport/handler/codec/CommonEncoder.java
@@ -9,8 +9,8 @@ import io.netty.handler.codec.MessageToByteEncoder;
 /**
  * <pre>
  *  -----------------------------------
- *  字节数 |   1   |    4    |    n   |
- *  解释  | 消息头 | 正文长度 |   正文  |
+ *  | 字节数 |   1   |    4    |    n   |
+ *  |  解释 | 消息头 | 正文长度 |   正文  |
  *  -----------------------------------
  * </pre>
  */
@@ -27,7 +27,7 @@ public class CommonEncoder extends MessageToByteEncoder<Message> {
     protected void encode(ChannelHandlerContext ctx, Message msg, ByteBuf out) throws Exception {
         int messageHeader = msg.messageType();
         byte[] data = serializationHandler.serialize(msg);
-        out.writeInt(messageHeader);
+        out.writeByte(messageHeader);
         out.writeInt(data.length);
         out.writeBytes(data);
     }
